v.2.3.0: see devel histories

+++ src/main/java/com/sx/datamarket/web/command/resource/DeleteDataTypesResourceCommand.java
@@ -1,33 +1,16 @@
 package com.sx.datamarket.web.command.resource;
 
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCResourceCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCResourceCommand;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextFactory;
-import com.liferay.portal.kernel.service.ServiceContextUtil;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.kernel.workflow.WorkflowConstants;
-import com.sx.icecap.constant.DataTypeProperties;
 import com.sx.icecap.constant.MVCCommand;
-import com.sx.icecap.constant.WebKey;
-import com.sx.constant.StationXConstants;
-import com.sx.constant.StationXWebKeys;
 import com.sx.icecap.constant.WebPortletKey;
-import com.sx.icecap.model.DataType;
 import com.sx.icecap.service.DataTypeLocalService;
-
-import java.io.PrintWriter;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Locale;
-import java.util.Map;
+import com.sx.util.portlet.SXPortletURLUtil;
 
 import javax.portlet.ResourceRequest;
 import javax.portlet.ResourceResponse;
@@ -54,21 +37,37 @@ public class DeleteDataTypesResourceCommand extends BaseMVCResourceCommand{
 		String strDataTypeIds = ParamUtil.getString(resourceRequest, "dataTypeIds", "[]");
 		System.out.println("strDataTypeIds: " + strDataTypeIds);
 		
-		JSONArray dataTypeIds = JSONFactoryUtil.createJSONArray(strDataTypeIds);
-		for(int i=0; i<dataTypeIds.length(); i++) {
-			long dataTypeId = dataTypeIds.getLong(i);
-			_dataTypeLocalService.removeDataType(dataTypeId);
-		}
-		
-		PrintWriter pw = resourceResponse.getWriter();
-		
 		JSONObject result = JSONFactoryUtil.createJSONObject();
 		
-		result.put("dataTypeIds", dataTypeIds);
+		if( strDataTypeIds.isEmpty() ) {
+			result.put("error", LanguageUtil.get(resourceRequest.getLocale(), "datatypes-should-be-specified"));
+			
+			SXPortletURLUtil.responeAjax(resourceResponse, result);
+			
+			return;
+		}
 		
-		pw.write(result.toJSONString());
-		pw.flush();
-		pw.close();
+		JSONArray errorObjects = JSONFactoryUtil.createJSONArray();
+		
+		String[] strDataTypeArray = strDataTypeIds.split(",");
+		for(String strDataTypeId : strDataTypeArray) {
+			try{ 
+				long dataTypeId = Long.parseLong(strDataTypeId);
+				
+				_dataTypeLocalService.removeDataType(dataTypeId);
+			} catch ( NumberFormatException e ) {
+				errorObjects.put( strDataTypeId );
+			}
+		}
+		
+		if( errorObjects.length() > 0 ) {
+			result.put("error", LanguageUtil.get(resourceRequest.getLocale(), "cannot-delete-some-of-datatypes"));
+			result.put("errorObjects", errorObjects);
+		}
+		
+		result.put("dataTypeList", strDataTypeIds);
+		
+		SXPortletURLUtil.responeAjax(resourceResponse, result);
 	}
 	
 	@Reference
